+++ src/util/stream.rs
@@ -1,3 +1,4 @@
+#[cfg(feature = "timer")]
 use tokio_timer::{
     throttle::Throttle,
     Timeout,
@@ -5,6 +6,7 @@ use tokio_timer::{
 
 use futures::Stream;
 
+#[cfg(feature = "timer")]
 use std::time::Duration;
 
 
@@ -25,6 +27,7 @@ pub trait StreamExt: Stream {
     /// Throttle down the stream by enforcing a fixed delay between items.
     ///
     /// Errors are also delayed.
+    #[cfg(feature = "timer")]
     fn throttle(self, duration: Duration) -> Throttle<Self>
     where Self: Sized
     {
@@ -63,6 +66,7 @@ pub trait StreamExt: Stream {
     /// tokio::run(stream);
     /// # }
     /// ```
+    #[cfg(feature = "timer")]
     fn timeout(self, timeout: Duration) -> Timeout<Self>
     where Self: Sized,
     {