+++ tokio-threadpool/src/builder.rs
@@ -3,6 +3,7 @@ use config::{Config, MAX_WORKERS};
 use park::{BoxPark, BoxedPark, DefaultPark};
 use shutdown::ShutdownTrigger;
 use pool::{Pool, MAX_BACKUP};
+use task::Queue;
 use thread_pool::ThreadPool;
 use worker::{self, Worker, WorkerId};
 
@@ -413,11 +414,13 @@ impl Builder {
             workers.into()
         };
 
+        let queue = Arc::new(Queue::new());
+
         // Create a trigger that will clean up resources on shutdown.
         //
         // The `Pool` contains a weak reference to it, while `Worker`s and the `ThreadPool` contain
         // strong references.
-        let trigger = Arc::new(ShutdownTrigger::new(workers.clone()));
+        let trigger = Arc::new(ShutdownTrigger::new(workers.clone(), queue.clone()));
 
         // Create the pool
         let pool = Arc::new(Pool::new(
@@ -425,6 +428,7 @@ impl Builder {
             Arc::downgrade(&trigger),
             self.max_blocking,
             self.config.clone(),
+            queue,
         ));
 
         ThreadPool::new2(pool, trigger)