+++ tokio-threadpool/src/lib.rs
@@ -1,4 +1,4 @@
-#![doc(html_root_url = "https://docs.rs/tokio-threadpool/0.1.9")]
+#![doc(html_root_url = "https://docs.rs/tokio-threadpool/0.1.10")]
 #![deny(warnings, missing_docs, missing_debug_implementations)]
 
 //! A work-stealing based thread pool for executing futures.
@@ -79,6 +79,7 @@
 
 extern crate tokio_executor;
 
+extern crate crossbeam_channel;
 extern crate crossbeam_deque as deque;
 extern crate crossbeam_utils;
 #[macro_use]