+++ tokio-threadpool/CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.1.10 (January 6, 2019)
+
+* Fix deadlock bug in `blocking` (#795).
+* Introduce global task queue (#798).
+* Use crossbeam's Parker / Unparker (#529).
+* Panic if worker thread cannot be spawned (#826).
+* Improve `blocking` API documentation (#789).
+
 # 0.1.9 (November 21, 2018)
 
 * Bump internal dependency versions (#746, #753).